+++ src/app/dashboard/account-card/account-card.component.css
@@ -1,5 +1,6 @@
 * {
   --error-clr: rgb(231, 73, 73);
+  --green-clr: rgb(85, 198, 113);
 }
 
 .eye {
@@ -47,6 +48,22 @@
   border: 4px solid var(--error-clr);
 }
 
+.add-span {
+  position: absolute;
+
+  top: 8px;
+}
+
+.green {
+  border: 4px solid var(--green-clr);
+}
+
+.green-btn {
+  border: 4px solid var(--green-clr);
+
+  color: var(--green-clr);
+}
+
 .half-col {
   height: 100%;
 
@@ -63,7 +80,7 @@
   margin-top: 20px;
 }
 
-@media screen and (min-width: 900px) {
+@media screen and (min-width: 700px) {
   .title {
     color: var(--grey-titles);
     width: 190px;
@@ -73,13 +90,16 @@
 
 .round-btn {
   background: white;
-  border: 4px solid var(--error-clr);
-
-  color: var(--error-clr);
 
   position: absolute;
   top: 35%;
   left: 45%;
+}
+
+.red-btn {
+  border: 4px solid var(--error-clr);
+
+  color: var(--error-clr);
 }
 
 .card-container {
@@ -104,10 +124,6 @@
   flex-direction: column;
   justify-content: space-around;
 }
-
-/* .half {
-  overflow-x: hidden;
-} */
 
 .card h3 {
   display: flex;
