+++ src/app/dashboard/actions/actions.component.css
@@ -20,4 +20,26 @@ a:hover {
 
 .panel {
   padding: 20px;
+
+  position: relative;
+}
+
+.not-active {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+
+  border-radius: 15px;
+
+  background-color: rgba(var(--menu-bg-rgb), 0.8);
+}
+
+.not-active span p {
+  padding: 20px;
+
+  color: white;
+  font-weight: bold;
+  text-shadow: 0 0 5px rgb(5, 14, 96);
 }
