+++ src/app/dashboard/actions/actions.component.html
@@ -7,7 +7,18 @@
 <div class="half">
   <div class="container-actions">
     <h3 class="title">Azioni</h3>
+
     <div class="panel">
+      <div class="not-active" *ngIf="!currentCard.active">
+        <span
+          ><p>
+            Nessuna azione è disponibile<br />
+            per questa carta inattiva. <br />Clicca il pulsante sulla carta per
+            aggiornare.
+          </p></span
+        >
+      </div>
+
       <a [routerLink]="['/dashboard']" [queryParams]="{ action: 'bonifico' }"
         ><h4><i class="fa-solid fa-dollar-sign"></i> Bonifico</h4></a
       >
