+++ src/app/dashboard/actions/pop-up/pop-up.component.css
@@ -50,6 +50,17 @@ select {
   z-index: 3;
 
   align-self: center;
+
+  position: relative;
+}
+
+.fa-circle-xmark {
+  position: absolute;
+  top: 10px;
+  right: 10px;
+  font-size: 20px;
+
+  cursor: pointer;
 }
 
 input,
@@ -81,13 +92,14 @@ label {
   margin: 0 5px 0 5px;
   font-size: 20px;
 
-  background: rgba(var(--menu-bg-rgb), 0.5);
+  background: rgba(var(--menu-font-hover-rgb), 0.05);
+  /* background: rgba(var(--menu-bg-rgb), 0.5);
   background: linear-gradient(
     90deg,
     rgba(var(--menu-bg-rgb), 0.5) 0%,
     rgba(var(--menu-font-hover-rgb), 0.5) 80%,
     rgba(var(--menu-font-hover-rgb), 0.5) 100%
-  );
+  ); */
 }
 
 .personalized-btn:active {
