+++ src/app/dashboard/dashboard.component.ts
@@ -1,5 +1,5 @@
 import { AuthService } from './../auth/auth.service';
-import { CardService } from './card-manage.service';
+import { CardService } from '../services/card-manage.service';
 import { Component, OnInit } from '@angular/core';
 
 @Component({
