+++ src/app/dashboard/graphic-area/graphic-area.component.html
@@ -1,7 +1,7 @@
 <div class="half">
   <div class="panel">
     <fusioncharts
-      [width]="width"
+      [width]="width.toString()"
       [height]="height"
       [type]="type"
       [dataFormat]="dataFormat"
