+++ src/app/dashboard/transactions/transactions.component.css
@@ -35,6 +35,8 @@ table {
 .panel {
   overflow-y: scroll;
   height: 220px;
+
+  width: 100%;
 }
 
 .personalized-btn {
@@ -44,12 +46,12 @@ table {
   color: var(--menu-bg-color);
   font-weight: bold;
 
-  background: rgba(var(--menu-bg-rgb), 0.5);
+  background: rgba(var(--menu-bg-rgb), 0.2);
   background: linear-gradient(
     90deg,
-    rgba(var(--menu-bg-rgb), 0.5) 0%,
-    rgba(var(--menu-font-hover-rgb), 0.5) 80%,
-    rgba(var(--menu-font-hover-rgb), 0.5) 100%
+    rgba(var(--menu-bg-rgb), 0.3) 0%,
+    rgba(var(--menu-font-hover-rgb), 0.2) 80%,
+    rgba(var(--menu-font-hover-rgb), 0.2) 100%
   );
 }
 
@@ -73,9 +75,17 @@ table {
   .container-actions {
     width: fit-content;
     margin: 0;
+
+    width: 90%;
   }
 
   .title {
     color: var(--grey-titles);
+  }
+}
+
+@media screen and (min-width: 900px) and (max-width: 1020px) {
+  .container-actions {
+    width: fit-content;
   }
 }
