+++ src/app/dashboard/transactions/transactions.component.ts
@@ -1,3 +1,4 @@
+import { CardService } from '../../services/card-manage.service';
 import { Component, OnInit } from '@angular/core';
 
 @Component({
@@ -6,143 +7,16 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./transactions.component.css'],
 })
 export class TransactionsComponent implements OnInit {
-  transactions = [
-    {
-      id: 1,
-      type: 'bonifico',
-      amount: 40,
-      date: '27-09-2021',
-      from: 'contoxxx',
-      to: 'contoyyy',
-      description: '1',
-    },
-    {
-      id: 1,
-      type: 'bonifico',
-      amount: 40,
-      date: '27-09-2021',
-      from: 'contoxxx',
-      to: 'contoyyy',
-      description: '2',
-    },
-    {
-      id: 1,
-      type: 'bonifico',
-      amount: 40,
-      date: '27-09-2021',
-      from: 'contoxxx',
-      to: 'contoyyy',
-      description: '3',
-    },
-    {
-      id: 1,
-      type: 'bonifico',
-      amount: 40,
-      date: '27-09-2021',
-      from: 'contoxxx',
-      to: 'contoyyy',
-      description: '4',
-    },
-    {
-      id: 1,
-      type: 'bonifico',
-      amount: 40,
-      date: '27-09-2021',
-      from: 'contoxxx',
-      to: 'contoyyy',
-      description: '5',
-    },
-    {
-      id: 1,
-      type: 'bonifico',
-      amount: 40,
-      date: '27-09-2021',
-      from: 'contoxxx',
-      to: 'contoyyy',
-      description: '6',
-    },
-    {
-      id: 1,
-      type: 'versamento',
-      amount: 3,
-      date: '27-09-2021',
-      from: 'contoxxx',
-      to: 'contoyyy',
-      description: '7',
-    },
-    {
-      id: 1,
-      type: 'versamento',
-      amount: 27.99,
-      date: '27-09-2021',
-      from: 'contoxxx',
-      to: 'contoyyy',
-      description: '8',
-    },
-    {
-      id: 1,
-      type: 'giroconto',
-      amount: 100,
-      date: '27-09-2021',
-      from: 'contoxxx',
-      to: 'contoyyy',
-      description: '9',
-    },
-    {
-      id: 1,
-      type: 'bonifico',
-      amount: 40.92,
-      date: '27-09-2021',
-      from: 'contoxxx',
-      to: 'contoyyy',
-      description: '10',
-    },
-    {
-      id: 1,
-      type: 'bonifico',
-      amount: 234,
-      date: '27-09-2021',
-      from: 'contoxxx',
-      to: 'contoyyy',
-      description: '11',
-    },
-    {
-      id: 1,
-      type: 'bonifico',
-      amount: 40,
-      date: '27-09-2021',
-      from: 'contoxxx',
-      to: 'contoyyy',
-      description: '12',
-    },
-    {
-      id: 1,
-      type: 'bonifico',
-      amount: 40,
-      date: '27-09-2021',
-      from: 'contoxxx',
-      to: 'contoyyy',
-      description: '13',
-    },
-    {
-      id: 1,
-      type: 'bonifico',
-      amount: 40,
-      date: '27-09-2021',
-      from: 'contoxxx',
-      to: 'contoyyy',
-      description: '14',
-    },
-    {
-      id: 1,
-      type: 'bonifico',
-      amount: 40,
-      date: '27-09-2021',
-      from: 'contoxxx',
-      to: 'contoyyy',
-      description: '15',
-    },
-  ];
+  transactions:
+    | {
+        id: number;
+        type: string;
+        amount: number;
+        date: string;
+        from: string;
+        to: string;
+        description: string;
+      }[] = [];
 
   transactionsDisplayed: {
     id: number;
@@ -154,9 +28,11 @@ export class TransactionsComponent implements OnInit {
     description: string;
   }[] = [];
 
-  constructor() {}
+  constructor(private cardService: CardService) {}
 
   ngOnInit(): void {
+    this.transactions = this.cardService.transactions;
+
     this.transactionsDisplayed = this.transactions;
   }
